change maxMismatch in PairedEndOverlap from absolute values to relative values

+++ MGX-Wizards/src/main/java/de/cebitec/mgx/gui/wizard/seqrun/SeqRunWizardIterator.java
@@ -285,8 +285,8 @@ public final class SeqRunWizardIterator implements WizardDescriptor.Iterator<Wiz
         return (Integer) wd.getProperty(SeqRunVisualPanel3.PROP_MINOVERLAP);
     }
     
-    public int getMaximalMismatches(){
-        return (Integer) wd.getProperty(SeqRunVisualPanel3.PROP_MAXMISMATCHES);
+    public float getMaximalMismatches(){
+        return (Float) wd.getProperty(SeqRunVisualPanel3.PROP_MAXMISMATCHES);
     }
